@@ -110,17 +117,15 @@
     "eslint-plugin-markdown": "^3.0.0",
     "eslint-plugin-no-template-curly-in-string-fix": "^1.0.4",
     "happy-dom": "^5.2.0",
-    "msw": "^0.48.2",
+    "msw": "^1.3.2",
     "node-html-parser": "^6.1.10",
     "npm-run-all": "^4.1.5",
     "nx": "15.0.10",
     "postcss": "^8.4.14",
     "prettier": "2.6.2",
     "prettier-plugin-tailwindcss": "^0.1.11",
-    "prisma": "^4.6.0",
     "tailwindcss": "^3.1.0",
-    "ts-node": "^10.8.1",
     "tsconfig-paths": "^4.0.0",
-    "typescript": "~4.8.2"
+    "typescript": "5.2.2"
   }
 }
