@@ -1,19 +1,20 @@
-const path = require("path")
-const { routeExtensions } = require("remix-custom-routes")
+import path from "path"
+import { routeExtensions } from "remix-custom-routes"
+
 /**
  * @type {import('@remix-run/dev').AppConfig}
  */
-module.exports = {
+export default {
   tailwind: true,
   postcss: true,
-  future: {
-    v2_routeConvention: true,
-    v2_meta: true,
-    v2_errorBoundary: true,
-  },
   cacheDirectory: "./node_modules/.cache/remix",
+  browserNodeBuiltinsPolyfill: {
+    modules: {
+      crypto: true,
+    },
+  },
   async routes() {
-    const appDirectory = path.join(__dirname, "app")
+    const appDirectory = path.join(process.cwd(), "app")
 
     return routeExtensions(appDirectory)
   },
