@@ -1,6 +1,6 @@
 import * as React from "react"
 
-import { cn } from "~/utils/misc"
+import { cn } from "~/utils/misc.ts"
 
 const Table = React.forwardRef<
   HTMLTableElement,
@@ -9,7 +9,7 @@ const Table = React.forwardRef<
   <div className="w-full overflow-auto">
     <table
       ref={ref}
-      className={cn("w-full caption-bottom text-sm", className)}
+      className={cn("caption-bottom w-full text-sm", className)}
       {...props}
     />
   </div>
@@ -42,7 +42,10 @@ const TableFooter = React.forwardRef<
 >(({ className, ...props }, ref) => (
   <tfoot
     ref={ref}
-    className={cn("bg-neutral-900 font-medium text-neutral-50 dark:bg-neutral-50 dark:text-neutral-900", className)}
+    className={cn(
+      "bg-neutral-900 font-medium text-neutral-50 dark:bg-neutral-50 dark:text-neutral-900",
+      className,
+    )}
     {...props}
   />
 ))
@@ -55,8 +58,8 @@ const TableRow = React.forwardRef<
   <tr
     ref={ref}
     className={cn(
-      "border-b transition-colors hover:bg-neutral-100/50 data-[state=selected]:bg-neutral-100 dark:hover:bg-neutral-800/50 dark:data-[state=selected]:bg-neutral-800",
-      className
+      "data-[state=selected]:bg-neutral-100 dark:data-[state=selected]:bg-neutral-800 border-b transition-colors hover:bg-neutral-100/50 dark:hover:bg-neutral-800/50",
+      className,
     )}
     {...props}
   />
@@ -70,8 +73,8 @@ const TableHead = React.forwardRef<
   <th
     ref={ref}
     className={cn(
-      "h-10 px-2 text-left align-middle font-medium text-neutral-500 [&:has([role=checkbox])]:pr-0 [&>[role=checkbox]]:translate-y-[2px] dark:text-neutral-400",
-      className
+      "[&:has([role=checkbox])]:pr-0 [&>[role=checkbox]]:translate-y-[2px] h-10 px-2 text-left align-middle font-medium text-neutral-500 dark:text-neutral-400",
+      className,
     )}
     {...props}
   />
@@ -85,8 +88,8 @@ const TableCell = React.forwardRef<
   <td
     ref={ref}
     className={cn(
-      "p-2 align-middle [&:has([role=checkbox])]:pr-0 [&>[role=checkbox]]:translate-y-[2px]",
-      className
+      "[&:has([role=checkbox])]:pr-0 [&>[role=checkbox]]:translate-y-[2px] p-2 align-middle",
+      className,
     )}
     {...props}
   />
@@ -99,7 +102,10 @@ const TableCaption = React.forwardRef<
 >(({ className, ...props }, ref) => (
   <caption
     ref={ref}
-    className={cn("mt-4 text-sm text-neutral-500 dark:text-neutral-400", className)}
+    className={cn(
+      "mt-4 text-sm text-neutral-500 dark:text-neutral-400",
+      className,
+    )}
     {...props}
   />
 ))
@@ -107,11 +113,11 @@ TableCaption.displayName = "TableCaption"
 
 export {
   Table,
-  TableHeader,
   TableBody,
+  TableCaption,
+  TableCell,
   TableFooter,
   TableHead,
+  TableHeader,
   TableRow,
-  TableCell,
-  TableCaption,
 }
