@@ -1,17 +1,17 @@
 // http://localhost:3000/content/react-headless-submenus/example
 
-import type { LoaderArgs } from "@remix-run/node"
+import type { LoaderFunctionArgs } from "@remix-run/node"
 import { json } from "@remix-run/node"
 import { useState } from "react"
 import {
-  Submenu,
   ButtonItem as ButtonItemCore,
-  SubmenuTrigger as SubmenuTriggerCore,
   MenuProvider,
-} from "../../components/Submenu"
-export { mergeHeaders as headers } from "~/utils/misc"
+  Submenu,
+  SubmenuTrigger as SubmenuTriggerCore,
+} from "../../components/Submenu.tsx"
+export { mergeHeaders as headers } from "~/utils/misc.ts"
 
-export async function loader({ request }: LoaderArgs) {
+export async function loader({ request }: LoaderFunctionArgs) {
   return json({})
 }
 
