@@ -1,5 +1,8 @@
-import path from "path"
 import fs from "fs"
+import path from "path"
+import { fileURLToPath } from "url"
+const __filename = fileURLToPath(import.meta.url)
+const __dirname = path.dirname(__filename)
 
 let redirects: Array<any> = []
 
