@@ -1,38 +1,21 @@
+import rehypeShiki from "@leafac/rehype-shiki"
 import { bundleMDX } from "mdx-bundler"
-import type { GitHubFile } from "~/types"
-import { getQueue } from "./p-queue.server"
+import { getQueue } from "./p-queue.server.ts"
+
+import chalk from "chalk"
+import remarkAutolinkHeadings from "remark-autolink-headings"
+import remarkGfm from "remark-gfm"
+import remarkSlug from "remark-slug"
+import * as shiki from "shiki"
+import { visit } from "unist-util-visit"
 
 async function compileMdxImpl<FrontmatterType extends Record<string, unknown>>({
   slug,
-  files,
+  content,
 }: {
   slug: string
-  files: Array<GitHubFile>
+  content: string
 }) {
-  // prettier-ignore
-  const { default: remarkAutolinkHeader } = await import("remark-autolink-headings")
-  const { default: remarkGfm } = await import("remark-gfm")
-  const { default: remarkSlug } = await import("remark-slug")
-
-  const { visit } = await import("unist-util-visit")
-
-  const indexPattern = /index.mdx?$/
-  const indexFile = files.find(({ path }) => path.match(indexPattern))
-  if (!indexFile) {
-    return null
-  }
-
-  const rootDir = indexFile.path.replace(indexPattern, "")
-  const relativeFiles = files.map(({ path, content }) => ({
-    path: path.replace(rootDir, "./"),
-    content,
-  }))
-
-  const filesObject = arrayToObject(relativeFiles, {
-    keyname: "path",
-    valuename: "content",
-  })
-
   const rehypeMetaAttribute = () => {
     return (tree) => {
       visit(tree, "element", visitor)
@@ -52,56 +35,47 @@ async function compileMdxImpl<FrontmatterType extends Record<string, unknown>>({
     }
   }
 
+  const highlighter = await shiki.getHighlighter({
+    theme: "poimandres",
+  })
   try {
-    const { code, frontmatter } = await bundleMDX({
-      source: indexFile.content,
-      files: filesObject,
+    console.log("Bundling MDX", slug)
+    const { code, frontmatter, errors } = await bundleMDX({
+      source: content,
       mdxOptions: (options) => ({
         remarkPlugins: [
           ...(options.remarkPlugins ?? []),
           remarkSlug,
-          [remarkAutolinkHeader, { behavior: "wrap" }],
+          [remarkAutolinkHeadings, { behavior: "wrap" }],
           remarkGfm,
         ],
         rehypePlugins: [
           ...(options.rehypePlugins ?? []),
           rehypeMetaAttribute,
           [
-            require(`rehype-shiki`),
+            rehypeShiki,
             {
-              theme: `public/from-paris-with-love.json`,
-              useBackground: true,
+              highlighter,
             },
           ],
         ],
       }),
     })
 
+    console.log({ errors })
+
     return { code, frontmatter: frontmatter as FrontmatterType }
   } catch (e) {
-    throw new Error(`MDX Compilation failed for ${slug}`)
+    console.error(chalk.red(`MDX Compilation failed for ${slug}`))
+  }
+
+  return {
+    code: null,
+    frontmatter: {} as Record<string, unknown>,
   }
 }
 
 const re = /\b([-\w]+(?![^{]*}))(?:=(?:"([^"]*)"|'([^']*)'|([^"'\s]+)))?/g
-
-function arrayToObject<Item extends Record<string, unknown>>(
-  array: Array<Item>,
-  { keyname, valuename }: { keyname: keyof Item; valuename: keyof Item },
-) {
-  const obj: Record<string, Item[keyof Item]> = {}
-
-  for (const item of array) {
-    const key = item[keyname]
-    if (typeof key !== "string") {
-      throw new Error(`Type of ${key} should be a string`)
-    }
-    const value = item[valuename]
-    obj[key] = value
-  }
-
-  return obj
-}
 
 async function queuedCompileMdx<
   FrontmatterType extends Record<string, unknown>,
